Windows support: fail to configure the WorkloadAttestor `unix` plugin
