+++ conf/agent/agent_full.conf
@@ -343,6 +343,7 @@ plugins {
 
     # WorkloadAttestor "unix": A workload attestor which generates unix-based
     # selectors like uid and gid.
+    # Supported on Unix only.
     WorkloadAttestor "unix" {
         plugin_data {
             # discover_workload_path: If true, the workload path will be discovered
