+++ doc/plugin_agent_workloadattestor_unix.md
@@ -56,3 +56,7 @@ A sample configuration:
 		}
 	}
 ```
+
+### Platform support
+
+This plugin is only supported on Unix systems.
